Remove watershed_processing.py and merge it to watershed.py
