+++ scripts/luna/watershed.py
@@ -1,7 +1,9 @@
-# Required Imports and loading up a scan for processing as presented by
-# Guide Zuidhof
+"""
+Watershed pre-processing.
+.. https://www.kaggle.com/ankasor/data-science-bowl-2017/improved-lung-segmentation-using-watershed
+"""
 
-
+import argparse
 import os
 
 import dicom
@@ -79,7 +81,7 @@ def generate_markers(image):
     return marker_internal, marker_external, marker_watershed
 
 
-def seperate_lungs(image):
+def separate_lungs(image):
     # Creation of the markers as shown above:
     marker_internal, marker_external, marker_watershed = generate_markers(
         image)
@@ -123,3 +125,39 @@ def seperate_lungs(image):
     segmented = np.where(lungfilter == 1, image, -2000 * np.ones((512, 512)))
 
     return segmented, lungfilter, outline, watershed, sobel_gradient, marker_internal, marker_external, marker_watershed
+
+
+def get_parser():
+    parser = argparse.ArgumentParser(description='watershed data')
+    parser.add_argument('-i', '--input-dir', nargs='?', required=True,
+                        help='input directory')
+    parser.add_argument('-o', '--output-path', nargs='?', required=True,
+                        help='output path directory')
+    return parser
+
+
+if __name__ == '__main__':
+    parser = get_parser()
+    args = parser.parse_args()
+
+    print(args)
+
+    input_dir = args.input_dir
+    output_path = args.output_path
+
+    scans = [dir_path for dir_path, _, _ in os.walk(input_dir)][1:]
+    for scan in scans:
+        p = scan.replace(input_dir, "")
+        print "processing " + p
+        patient_scan = load_scan(scan)
+        imgs = get_pixels_hu(patient_scan)
+        num_images = imgs.shape[0]
+        out_imgs = np.ndarray([num_images, 1, 512, 512], dtype=np.float32)
+        i = 0
+        for img in imgs:
+            if i % 10 == 0:
+                print 'imgage ' + str(i)
+            segmented, _, _, _, _, _, _, _ = separate_lungs(img)
+            out_imgs[i, 0, :, :] = segmented
+            i += 1
+        np.save(output_path + p + ".npy", out_imgs)
